take-home/BE/pyproject.toml
[project]
name = "take-home"
version = "0.2.0"
description = "CLI chat agent built on Claude Agent SDK with tool-based memory strategies"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk",
    "anthropic",
    "langchain",
    "langchain-anthropic",
    "python-dotenv",
    "sentence-transformers",
    "numpy",
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "deepagents>=0.4.5",
]

[project.scripts]
chat = "agent.cli:main"

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
testpaths = ["evals"]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]
